chore(home): move useScrollAnimation to a composable
