+++ pages/index.vue
@@ -19,32 +19,6 @@ ScrollTrigger.defaults({
 
 gsap.registerPlugin(ScrollTrigger)
 
-function useScrollAnimation(trigger: string, xPercent: number) {
-  const timeline = gsap.timeline({
-    scrollTrigger: {
-      trigger: trigger,
-      start: 'top bottom',
-      end: 'bottom bottom',
-      scrub: true,
-    },
-    defaults: {
-      ease: 'Power3.easeIn',
-    },
-  })
-
-  timeline.fromTo(
-    trigger,
-    {
-      opacity: 0,
-      xPercent: xPercent,
-    },
-    {
-      opacity: 1,
-      xPercent: 0,
-    }
-  )
-}
-
 onMounted(() => {
   const introParallax = gsap.timeline()
